Add support for Bottlerocket
Bottlerocket is an AWS operating system focused on providing a secure
environment for container workloads. It has SELinux enabled, so some
SELinux configuration was necessary. The installer was rewritten in go
to avoid ballooning dependencies with SELinux.

In Bottlerocket a container cannot be both privileged and have
SELinux super_t. SELinux is required to communicate with systemd and
privileged is required to get mount information and unmount. To work
around this limitation, umount now goes through systemd as well and
mount information is read from a mounted volume.