Clean up tokens by parsing target path in `NodeUnpublishVolume` (#251)
* Add `ParseTargetPath` function

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>

* Clean up tokens by parsing target path in `NodeUnpublishVolume`

We need Pod ID and Volume ID to clean up any previously created
tokens. `NodeUnpublishVolume` call receives Volume ID but not Pod ID.
We were keeping the mapping from target path to Pod ID in memory
before but this is not ideal as it's not persistent and we were losing
this data when CSI driver restarts. With this change, we're parsing
target path to get Pod ID and use that to clean up tokens without
storing any state in the CSI driver.

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>

* Use a RegExp to parse target path

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>

* Update target path RegExp

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>

---------

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>