Move testing S3 client to its own package (#294)
This PR moves S3 client used in tests to its own package for making it
reusable. It also changes the logic for generating testing bucket names.
Previously we were using cluster name as prefix, and also we had a logic
to truncate bucket name to make space for `--azid--x-s3` suffix. That
was causing unique hash in the bucket name to be truncated and different
clusters to use the same name, and it was causing spurious failures such
as
https://github.com/awslabs/mountpoint-s3-csi-driver/actions/runs/11835794931/job/32979499659#step:10:2599.

---

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>