Add `aws-s3-csi-mounter` component (#291)
This is part of
https://github.com/awslabs/mountpoint-s3-csi-driver/issues/279.

This new component, `aws-s3-csi-mounter`, will be the entry point for
the container running Mountpoint. It will be responsible for receiving
mount options and FUSE file descriptor and spawning Mountpoint process
until completion.

---

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>