Add `aws-s3-csi-controller` component (#299)
This is part of
https://github.com/awslabs/mountpoint-s3-csi-driver/issues/279.

This new component, `aws-s3-csi-controller`, will be the entry point for
our controller component. It's using
[controller-runtime](https://github.com/kubernetes-sigs/controller-runtime),
specifically, it implements
[`Reconciler`](https://pkg.go.dev/sigs.k8s.io/controller-runtime/pkg#hdr-Reconciler)
interface to reconcile Pods in the cluster. It schedules Mountpoint Pods
in turn to cluster events such as a new workload Pod using a PV backed
by S3 CSI Driver getting scheduled into the cluster. It'd then schedule
a Mountpoint Pod for that workload Pod in the same node to provide
volume for that Pod.

https://github.com/awslabs/mountpoint-s3-csi-driver/issues/279 is still
WIP and this component contains some TODOs and it's not in use anywhere
except in tests at the moment.

---

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>