Stop retrying to delete security group if we got "NotFound" error (#324)
We were missing "NotFound" checks in our CI while trying to remove
security groups, and it was causing this operation to retry until max
retry time which is 10 minutes. This was resulting temporary credentials
issued for CI to expire, and the CI run to fail. See
https://github.com/awslabs/mountpoint-s3-csi-driver/actions/runs/12392057877/job/34648273564
for example. We might still need to increase duration of temporary
credentials from 1 hour to a longer duration if our CI keeps taking
longer, but for now this should address the issue, hopefully.

---

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>