Update CI to use Github Environments (#326)
*Issue #, if available:* N/A

*Description of changes:*

Use Github environments to separate the different security boundaries we
have: an 'untrusted' account for third party forks, and a 'trusted'
account for direct pushes and merges. The 'trusted' account is then used
as a source for releases, and it will only run code that's already
merged into the main repository.

Testing E2E:
https://github.com/awslabs/mountpoint-s3-csi-driver/actions/runs/12653780902
Testing release:
https://github.com/muddyfish/mountpoint-s3-csi-driver/actions/runs/12411681334

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.