Create a shorter symlink for Unix socket in `mountoptions` package
Go returns `invalid argument` if you try to `net.{Dial, Listen}` on a
Unix socket if the path is longer than 108 characters.

This limitation is only to the argument passed to those functions, if
we receive a Unix socket path longer than 108 characters, we create a
symlink for that path in tempdir for that platform and use that path
in network operations to workaround this limitation.

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>