Add approval gate environment (#339)
*Issue #, if available:* N/A

*Description of changes:* Adds an 'approval gate' step, which only waits
to be approved. In the main MP CSI Driver repo, added an environment
'approval-gate' which needs an approval to run. Adding the 'needs'
delays the rest of the tests from running until an admin approves, and
means we can remove the approver from the 'untrusted' environment after
merging


Closes https://github.com/awslabs/mountpoint-s3-csi-driver/pull/338, as
previous PR was on a 'workflow' branch, and this is not required and
increases the time taken to merge in.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.