Write error log to `mount.err` if Mountpoint fails to start (#347)
This is to let `PodMounter` doing this mount operation know that mount
failed immediately instead of waiting for timeout, and also to propagate
mount errors to users.

---

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>