Allow overriding of hardcoded DaemonSet tolerations (#354)
Resolves #348 

*Description of changes:*

This PR will move the hardcoded tolerations from the DaemonSet template
to the values, ~~I just used the
[aws-fsx-csi-driver](https://github.com/kubernetes-sigs/aws-fsx-csi-driver/blob/master/charts/aws-fsx-csi-driver/templates/node-daemonset.yaml#L39-L45)
helm chart as inspiration. As this is probably the simplest solution, it
may lead to unexpected changes if someone already has `tolerations`
set.~~
I have added a new value: `node.defaultTolerations` and put it in the
`else` condition in `tolerateAllTaints`.

---

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.