Use relative paths for Unix sockets in `mountoptions` package
Go returns `invalid argument` if you try to `net.{Dial, Listen}` on a
Unix socket if the path is longer than 108 characters.

This limitation is only to the argument passed to those functions, if
we receive a Unix socket path longer than 108 characters, we try to
make that path relative to the current working directory and use it
instead if relative path is shorter than 108 characters.

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>