Add context argument to `mounter.Unmount` method (#371)
`SystemdMounter` was creating a background context in the constructor
and using it when a context needed – which is not a good practice as
main reason for contexts are cancellation and timeouts but creating a
background context kills that benefit. Now, `mounter.Unmount` accepts
context as a first argument.

---

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.