Use a tempdir for cache if caching is enabled in `csimounter` (#379)
This is a temporary workaround in `csimounter` to create a cache folder
in caching is enabled. We should ideally create a volume in the
Mountpoint Pod and use that as cache, but until we implement that, this
workaround allows us to pass end-to-end tests that rely on cache.

---

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.