EKS Pod Identity support for driver level credentials (#451)
This PR is a follow-up to
https://github.com/awslabs/mountpoint-s3-csi-driver/pull/449 because we
need the E2E tests to run with the newly added `cluster-name` parameter.
This is not possible in a PR created from a fork.

*Description of changes:* With this change, we will support the
configuration of driver level credentials with EKS Pod Identity. In EKS
clusters, this is an alternative to IRSA with an easier configuration
process.

*Key Changes:*

- Credential Provider
- In `provider_driver.go` the container credentials used for EKS Pod
Identity are fetched from the driver pod and forwarded to the Mountpoint
process. This is done similarly to STS Web Identity credentials used for
IRSA.
- Added unit tests covering different credential setup scenarios
(including combinations)

- Testing
  - Added cluster name as parameter for E2E tests
- Added E2e tests covering EKS Pod Identity setup scenarios with IAM
roles of varying levels of S3 access
  - Manual testing performed on personal EKS cluster

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Signed-off-by: Renan Magagnin <renanmag@amazon.co.uk>