EKS Pod Identity support for pod level credentials (#458)
*Description of changes:* With this change, we will support the
configuration of pod level credentials with EKS Pod Identity. In EKS
clusters, this is an alternative to IRSA with an easier configuration
process.

*Key Changes:*

- Credential Provider
- In `provider_pod.go` the precedence is given to IRSA credentials. If
they are not configured, then EKS Pod Identity credentials are forwarded
to the Mountpoint process instead.
  - Added unit tests

- Testing
- Added E2E tests covering EKS Pod Identity setup scenarios with IAM
roles of varying levels of S3 access
- Added one pod level IRSA E2E test to make sure driver level EKS Pod
Identity is not used instead
  - Added E2E tests covering setup scenario of IRSA and EKS Pod Identity
  - Manual testing performed on personal EKS cluster

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Signed-off-by: Renan Magagnin <renanmag@amazon.co.uk>