Do not mount `plugin-dir` in CSI Driver container
The plugin directory (`/var/lib/kubelet/plugins/s3.csi.aws.com/` by default) was mounted in the CSI Driver container at `/csi`.
However, the kubelet directory (`/var/lib/kubelet` by default) is also mounted in the CSI Driver at `/var/lib/kubelet` which already includes the plugin dir.
This was okay so far, but with Pod sharing we started mounting Mountpoint instances in the plugin dir, and also bind mounts from the kubelet dir for the workloads.
This was causing Mountpoint mounts to have references to themselves in both the plugin dir and the kubelet dir, and due that when we try to unmount them,
we were getting "resource is busy" errors, as the kernel was thinking there are still references to the mount.

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>