S3CSI-69: Update supported K8s version to >1.30
Updated the supported k8s version to 1.30.0-0 and above in the helm
chart as thats what is needed by our customers. We are intentionally
reducing the scope to make maintainability easier. We currently have no
need for supporting older versions of k8s. This is a change in our
supported versions and not a change in the minimum required version.