Improve logging in `PodUnmounter` (#480)
`PodUnmounter` is idempotent and might be called multiple times for the
same operation due to way it works, up to this PR we were logging stuff
in each invocation which was a bit noisy. Now, we only log if we really
performed some operations.

---

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>