Remove kOps testing clusters (#528)
We were having problems on updating kOps version to latest, and was
using an old version of kOps. Because of this, we were only able to use
Kubernetes versions up to 1.28. Given that, and we're [getting some
transient
errors](https://github.com/awslabs/mountpoint-s3-csi-driver/issues/288)
with kOps clusters time to time, we decided to remove kOps clusters in
our tests for now. We can add them back properly using the latest
version and other Kubernetes versions in the future if we'd like to.

---

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>