Add end-to-end test for upgrading the CSI Driver (#532)
## Change

This new test suite verifies that upgrading the CSI Driver to a new
version doesn't break existing workloads that created with the previous
version. The test involves the following steps for testing the upgrade
from X to Y:
- Uninstall the CSI Driver if it's installed with any version
- Install version X
- Start three workloads
   - A Pod using `S3ReadOnlyAccess` policy with driver-level IRSA
   - A Pod using `S3FullAccess` policy with pod-level IRSA
   - A Pod using `S3ReadOnlyAccess` policy with pod-level IRSA
- Write a test file on writable workloads, and assert read-only
workloads can list but cannot write
- Upgrade to version Y
- Start three workloads
   - A Pod using `S3ReadOnlyAccess` policy with driver-level IRSA
   - A Pod using `S3FullAccess` policy with pod-level IRSA
   - A Pod using `S3ReadOnlyAccess` policy with pod-level IRSA
- Write a test file on writable workloads, and assert read-only
workloads can list but cannot write
- Every minute for an hour repeat:
- Ensure all writable pods can read the original file, and can do basic
file operations (i.e., writing, reading, listing and deleting)
   - Ensure all read-only pods can still list but cannot write
- Ensure all workloads can be successfully terminated

## Testing

Verified that this test fails if I disable `systemd` mount support by
setting `supportLegacySystemDMounts` to `false`. This might happen after
2 hours or so because:
1. We ask for service account tokens that valid for 1 hour (see
`CSIDriver` object)
  2. Session duration of the IAM roles we assume is 1 hour

So, it might take ~2 hours at worst case to hit both cycles, and get the
failure. That's why we run the upgrade test for 150 minutes.

## TODOs:

- [x] Add a GitHub workflow to manually trigger upgrade tests between
two given versions - Added a workflow with
https://github.com/awslabs/mountpoint-s3-csi-driver/pull/532/commits/2348101b9c99567904bd80d63fa69c342ff68d83
but couldn't test it. It needs to be in main to be able to triggered
manually, so, I'll send follow-up PRs if I made a mistake on the
workflow file – which is very likely 😅

---

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>