Increase max retry attempts and backoff in end-to-end tests
With a recent addition to our testing matrix, we now run our end-to-end tests with 44 clusters in parallel.
The tests in those clusters are also run in parallel and this caused some throttling errors in various AWS services especially from the IAM service.

This change tries to make the test suite more resilient by increasing the maximum retry amount from 3 to 5 and also increasing maximum backoff duration from 20s to 2m.

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>