Fix broken pod sharing link (#582)
*Description of changes:*

The README for the static provisioning example contained a relative link
to a section of a non-existent document,
`../../docs/FEATURES.md#pod-sharing`. This contribution corrects that
link to `../../../docs/MOUNTPOINT_POD_SHARING.md`.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.