Add env var for EKS Pod Identity token audience and service principal  (#617)
*Issue #, if available:* N/A

*Description of changes:*

For testing in internal EKS environments, we may wish to modify the
token audience and service principal used. This change allows it to be
overridden in the node service and end-to-end tests.

The default value is unchanged, and remains `"pods.eks.amazonaws.com"`.

We may consider making changes to the Helm chart, however I'm leaving it
out for this PR.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.