Add env var for EKS Pod Identity token audience and service principal  (#617)
*Issue #, if available:* N/A

*Description of changes:*

For testing in internal EKS environments, we may wish to modify the
token audience and service principal used. This change allows it to be
overridden in the node service and end-to-end tests.

The default value is unchanged, and remains `"pods.eks.amazonaws.com"`.

We may consider making changes to the Helm chart, however I'm leaving it
out for this PR.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Remove overrides for EKS PLI token audience, keep for AWS service principal (#622)

*Issue #, if available:* N/A

*Description of changes:*

The project previously conflated JWT token audiences and AWS service
principals, hence an earlier change was not quite the right one:
https://github.com/awslabs/mountpoint-s3-csi-driver/commit/2660105356d03e056f0126d4aafc7b8d197cdaa0

This change partially reverts that, and introduces configuration for the
AWS service principal only.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>