Add explicit permissions to all GHA workflows (#630)
*Issue #, if available:* N/A

*Description of changes:*

Adding explicit permissions to adopt least-privilege principle in
workflows.

Some permissions were no longer used: for example, we no longer publish
the test coverage although we still tried to write it to the job
summary. This removes that for now.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>