Update node daemonset to use host PID namespace (#632)
*Issue #, if available:* #626

*Description of changes:*

With Mountpoint CSI Driver V2, the opening of the FUSE device moved from
the invocation of Mountpoint as a host process to the node daemon which
now hands it over to an unprivileged Mountpoint. Unfortunately, this
changed the PID namespace from the host namespace to a namespace
specific to the node daemon impacting how the kernel sends PIDs to
Mountpoint. All PIDs result in zero, as the FUSE daemon should not have
visibility into other namespaces according to the kernel.

This change moves the node daemon to the host PID namespace. The FUSE
device will now be opened in the host PID namespace, and the FUSE driver
in the kernel will hand over host PIDs to Mountpoint allowing it to
correctly implement Mountpoint semantics.

The change has been tested manually by rerunning an application that was
impacted by this bug, and it has additionally been tested by installing
in an ROSA cluster and verifying pods can be started (but no
verification of Mountpoint behavior itself).

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>