Add wait flag to end-to-end test eksctl delete usage, and disable safety checks (#639)
*Issue #, if available:* N/A

*Description of changes:*

This change updates the invocation of `eksctl` CLI to correctly wait for
the deletion of the cluster. Additionally, I disable some eviction
safety checks since we do not care to ensure that the cluster is torn
down 'safely' since this is just for end-to-end tests.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>