Move Helm release to manually dispatched workflow (#642)
*Issue #, if available:* N/A

*Description of changes:*

There's currently a small window between where the charts are published
and when the images are published. This could mean there's a window of
time where installs of the driver will fail to bring up components in
clusters.

Move Helm release to a manually dispatched workflow for now.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>