Add post_test CI target for controller tests, set unit test workflow name (#641)
*Issue #, if available:* N/A

*Description of changes:*

Add a post_test target so we can set a requires relationship for merges.
Additionally, update the name of the unit test workflow since "build"
doesn't make sense.

After this is merged, we should update the required targets in the
repository configuration.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>