Add documentation for Mountpoint metrics (#629)
*Issue #, if available:* N/A

*Description of changes:*

This introduces initial documentation showing how to get started with
Mountpoint metrics using the CloudWatch Observability add-on.

We're not prescribing any particular solution here. If we wanted to
extend to ADOT later, for example, that would be a nice improvement.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>