Update Helm-chart to disable hostPID on OpenShift, add isOpenShift Helm value, disable hostPID on Kustomize (#667)
*Issue #, if available:* #655

*Description of changes:*

This change rolls back the hostPID fix in #632 for OpenShift clusters
only, and only where Helm chart is the installation method. Kustomize
installs are included in the roll back.

This means that OpenShift clusters will continue to be impacted by
https://github.com/awslabs/mountpoint-s3-csi-driver/issues/626.

Manually tested on a ROSA cluster, verifying that files can be written
to the mounted bucket.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>